refactor: member의 name 원시타입 포장
