+++ README.md
@@ -30,3 +30,5 @@
  
 - [ ] 개인별 이름을 입력하면 개인별 결과를 출력
 - [ ] "all"을 입력하면 전체 참여자의 실행 결과를 출력한다.
+
+- [x] member의 name 원시타입 포장
