+++ src/main/java/nextstep/ladder/domain/Member.java
@@ -1,34 +1,15 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.exceptions.CannotNullOrEmptyException;
-import nextstep.ladder.domain.exceptions.ExceedMaxNameLengthException;
-
 public class Member {
-    public static final int MAX_NAME_LENGTH = 5;
-
-    private final String name;
+    private final Name name;
 
     public Member(String name) {
-        checkNullOrBlank(name);
-        checkExceedMaxLength(name);
-        this.name = name;
-    }
-
-    private void checkNullOrBlank(String name) {
-        if (name == null || name.isEmpty()) {
-            throw new CannotNullOrEmptyException();
-        }
-    }
-
-    private void checkExceedMaxLength(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new ExceedMaxNameLengthException();
-        }
+        this.name = new Name(name);
     }
 
     @Override
     public String toString() {
-        return this.name;
+        return this.name.toString();
     }
 }
