+++ src/main/java/nextstep/ladder/domain/Members.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Members {
-    private  static final String DELIMITER = " ";
-    private final static String NAME_DELIMITER = ",";
+    private static final String DELIMITER = " ";
+    private static final String NAME_DELIMITER = ",";
 
     private final List<Member> members;
 
@@ -35,7 +35,7 @@ public class Members {
     public String getPrettyString() {
         return members.stream()
                 .map(Member::toString)
-                .map(name -> name + DELIMITER.repeat(Member.MAX_NAME_LENGTH - name.length()))
+                .map(name -> name + DELIMITER.repeat(Name.MAX_NAME_LENGTH - name.length()))
                 .collect(Collectors.joining(DELIMITER))
                 .strip();
     }
